Moving tests to new file
